pyproject.toml
[project]
name = "stavki"
version = "2.0.0"
description = "Sports Betting Analytics System with ML"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "STAVKI Team"}
]

dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scipy>=1.9.0",
    "scikit-learn>=1.1.0",
    "catboost>=1.1.0",
    "lightgbm>=3.3.0",
    "torch>=1.12.0",
    "requests>=2.28.0",
    "python-dotenv>=0.21.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
]
telegram = [
    "python-telegram-bot>=20.0",
]
all = [
    "stavki[dev,telegram]",
]

[project.scripts]
stavki = "stavki.interfaces.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["stavki*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
target-version = "py39"
